+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-# CHANGELOG for v0.1.x
+# CHANGELOG for v0.x
 
-## v0.1.1 (2022-10-06)
+## v0.2.0 (2022-11-19)
+### Enhancements
+* Moved from inspecting query on tests. Added sample schemas and repo
+* Added sorting by joint query `EctoJuno.Query.Sorting.sort_query/4`
+
+## v0.1.1 (2022-11-06)
 ### Enhancements
 * Moved to `default` option for validator schema default sorting parameters
  
@@ -8,7 +13,7 @@
 * Fixed mandatory `params` argument in `EctoJuno.Query.Sorting.sort_query/3`
 * Fixed typos in `README.md` and other docs
 
-## v0.1.0 (2022-10-05)
+## v0.1.0 (2022-11-05)
 ### Enhancements
 * Added parameter parser and validator
 * Added sorting module for query